Remove replace subnet dependency
