+++ rs/ic-management-backend/src/endpoints/release.rs
@@ -18,9 +18,3 @@ pub(crate) async fn blessed(request: web::Path<ReleaseRequest>, registry: web::D
     let registry = registry.read().await;
     response_from_result(registry.blessed_versions(&request.release_artifact).await)
 }
-
-#[get("/release/versions/nns")]
-pub(crate) async fn get_nns_replica_version(registry: web::Data<Arc<RwLock<RegistryState>>>) -> Result<HttpResponse, Error> {
-    let registry = registry.read().await;
-    Ok(HttpResponse::Ok().json(registry.nns_replica_version().await))
-}
